refactor: correct headers
